app/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai>=1.3.0
numpy>=1.24.3
python-dotenv==1.0.0
pandas>=2.0.3
scikit-learn>=1.3.0
python-pptx>=0.6.21
Pillow>=10.2.0
matplotlib>=3.8.0
PyPDF2>=3.0.0

